+++ app/controllers/Gestores.java
@@ -1,22 +1,32 @@
 package controllers;
 
+import models.Area;
 import models.Gestor;
 import play.data.Form;
 import play.mvc.Controller;
 import play.mvc.Result;
 
-import com.avaje.ebean.Page;
-
 public class Gestores extends Controller{
 
 	private static final Form<Gestor> gestorForm = Form.form(Gestor.class);
+	public static Result GO_HOME = redirect(
+	        routes.Gestores.list(0, "nome", "asc", "")
+	    );
+	public static Result list(int page, String sortBy, String order, String filter) {
+        return ok(
+        		views.html.gestores.list.render(
+				Gestor.page(page, 10, sortBy, order, filter),
+                sortBy, order, filter
+            )
+        );
+    }
 	public static Result index() {
-		return redirect(routes.Gestores.lista(0));
-	  }
-	public static Result lista(Integer page){
+        return GO_HOME;
+    }
+	/*public static Result lista(Integer page){
 		Page<Gestor> gestores = Gestor.buscarTodos(page);
 		return ok(views.html.gestores.list.render(gestores));
-	}
+	}*/
 	public static Result novo(){
 		return ok(views.html.gestores.detalhes.render(gestorForm));
 	}
@@ -38,14 +48,18 @@ public class Gestores extends Controller{
 	    }
 		flash("success",
 		        String.format("Gestor atualizado %s", gestor));
-		return redirect(routes.Gestores.lista(0));
+		return redirect(routes.Gestores.list(0, "nome", "asc", ""));
 	}
 	public static Result delete(Long id) {
 	  final Gestor gestor = Gestor.buscarPorId(id);
 	  if(gestor == null) {
 		  return ok();
 	  }
+	  final Area area = Area.buscarPorGestor(id);
+	  if(area != null) {
+	    return badRequest(String.format("Gestor %s não pode ser excluído.", id));
+	  }
 	  gestor.delete();
-	  return redirect(routes.Gestores.lista(0));
+	  return redirect(routes.Gestores.list(0, "nome", "asc", ""));
 	}
 }
